Convert model-helpers to Typescript
