+++ imports/lib/models/users.d.ts
@@ -1,0 +1,11 @@
+import { Meteor } from 'meteor/meteor';
+
+declare module 'meteor/meteor' {
+  module Meteor {
+    interface User {
+      lastLogin?: Date;
+      roles?: string[];
+      hunts: string[];
+    }
+  }
+}
