+++ package.json
@@ -50,6 +50,7 @@
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.1.0",
     "@types/meteor": "^1.4.24",
+    "@types/mustache": "^0.8.32",
     "assert": "^1.4.1",
     "babel-eslint": "git+https://github.com/babel/babel-eslint.git",
     "babel-preset-meteor": "^7.2.0",
